+++ app/src/tab/live_tv.cpp
@@ -6,6 +6,7 @@
 #include "api/jellyfin.hpp"
 #include "view/video_card.hpp"
 #include "activity/player_view.hpp"
+#include "utils/keybind.hpp"
 #include <fmt/format.h>
 
 using namespace brls::literals;  // for _i18n
@@ -55,6 +56,13 @@ LiveTV::LiveTV(const std::string& itemId) {
     brls::Logger::debug("LiveTV: create {}", itemId);
 
     this->registerAction("hints/refresh"_i18n, brls::BUTTON_BACK, [this](...) {
+        this->recycler->showSkeleton();
+        this->doRequest();
+        return true;
+    });
+
+    this->registerAction(KeyBind::getRefresh(), [this](...) {
+        this->recycler->showSkeleton();
         this->doRequest();
         return true;
     });