+++ app/src/utils/config.cpp
@@ -53,6 +53,7 @@ namespace fs = std::experimental::filesystem;
 #include <borealis/views/edit_text_dialog.hpp>
 #include "api/jellyfin.hpp"
 #include "utils/config.hpp"
+#include "utils/keybind.hpp"
 #include "utils/misc.hpp"
 #include "utils/ums.hpp"
 #include "utils/thread.hpp"
@@ -131,6 +132,20 @@ std::unordered_map<AppConfig::Item, AppConfig::Option> AppConfig::settingMap = {
     {REQUEST_TIMEOUT, {"request_timeout", {"1000", "2000", "3000", "5000"}, {1000, 2000, 3000, 5000}}},
     {HTTP_PROXY_STATUS, {"http_proxy_status"}},
     {HTTP_PROXY, {"http_proxy"}},
+
+    {KEY_REFRESH, {"key_refresh"}},
+    {KEY_LAST, {"key_last"}},
+    {KEY_NEXT, {"key_next"}},
+    {KEY_VOLUME_UP, {"key_volume_up"}},
+    {KEY_VOLUME_DOWN, {"key_volume_down"}},
+    {KEY_VIDEO_PROFILE, {"key_video_profile"}},
+    {KEY_DANMAKU, {"key_danmaku"}},
+    {KEY_FORWARD, {"key_forward"}},
+    {KEY_REWIND, {"key_rewind"}},
+    {KEY_SETTING, {"key_setting"}},
+    {KEY_VIDEO_QUALITY, {"key_video_quality"}},
+    {KEY_VIDEO_SPEED, {"key_video_speed"}},
+    {KEY_VIDEO_PAUSE, {"key_video_pause"}},
 };
 
 static std::string generateDeviceId() {
@@ -321,6 +336,22 @@ bool AppConfig::init() {
     brls::Application::setFPSStatus(this->getItem(SHOW_FPS, false));
     VideoContext::swapInterval = this->getItem(SWAP_INTERVAL, 1);
 
+    // 初始化 KeyBind
+    KeyBind::setLast(this->getItem(KEY_LAST, std::string{"pgup"}));
+    KeyBind::setNext(this->getItem(KEY_NEXT, std::string{"pgdn"}));
+    KeyBind::setVolumeUp(this->getItem(KEY_VOLUME_UP, std::string{"0"}));
+    KeyBind::setVolumeDown(this->getItem(KEY_VOLUME_DOWN, std::string{"9"}));
+    KeyBind::setDanmaku(this->getItem(KEY_DANMAKU, std::string{"d"}));
+    KeyBind::setVideoProfile(this->getItem(KEY_VIDEO_PROFILE, std::string{"f1"}));
+    KeyBind::setVideoQuality(this->getItem(KEY_VIDEO_QUALITY, std::string{"f2"}));
+    KeyBind::setVideoSpeed(this->getItem(KEY_VIDEO_SPEED, std::string{"f3"}));
+    KeyBind::setSetting(this->getItem(KEY_SETTING, std::string{"f4"}));
+    KeyBind::setRefresh(this->getItem(KEY_REFRESH, std::string{"f5"}));
+    KeyBind::setForward(this->getItem(KEY_FORWARD, std::string{"]"}));
+    KeyBind::setRewind(this->getItem(KEY_REWIND, std::string{"["}));
+    KeyBind::setVideoOsd(this->getItem(KEY_VIDEO_OSD, std::string{"o"}));
+    KeyBind::setVideoPause(this->getItem(KEY_VIDEO_PAUSE, std::string{"space"}));
+
     // 初始化一些在创建窗口之后才能初始化的内容
     brls::Application::getWindowCreationDoneEvent()->subscribe([this]() {
         // 是否交换按键
@@ -371,24 +402,14 @@ bool AppConfig::init() {
         brls::Application::getPlatform()->getInputManager()->getKeyboardKeyStateChanged()->subscribe(
             [this](brls::KeyState state) {
                 if (!state.pressed) return;
-                auto top = brls::Application::getActivitiesStack().back();
                 switch (state.key) {
-                case brls::BRLS_KBD_KEY_F:
-                    // 在编辑框弹出时不触发
-                    if (dynamic_cast<brls::EditTextDialog*>(top->getContentView())) break;
 #ifndef __APPLE__
                 case brls::BRLS_KBD_KEY_F11:
-#endif
                     VideoContext::FULLSCREEN = !this->getItem(AppConfig::FULLSCREEN, VideoContext::FULLSCREEN);
                     this->setItem(AppConfig::FULLSCREEN, VideoContext::FULLSCREEN);
                     brls::Application::getPlatform()->getVideoContext()->fullScreen(VideoContext::FULLSCREEN);
                     break;
-                case brls::BRLS_KBD_KEY_SPACE: {
-                    MPVCore::instance().togglePlay();
-                    VideoView* video = dynamic_cast<VideoView*>(top->getContentView()->getView("video"));
-                    if (video) video->showOSD(true);
-                    break;
-                }
+#endif
                 default:;
                 }
             });